Works perfectly, and makefile.
What wasn't working was when the tree and text encoded was 0s and 1s.
And it was usually not divisible by 8. That resulted in trailing 0s that
were not a part of the message.

I thought of a solution that would be adding bits at the front to make
the total count divisible by 8, resulting in no trailing 0s.

The format was heavily, and I mean HEAVILY inspired by UTF8 encoding (because it's brilliant), it appends N-1 1s and a
0, at the start, then the other bits. There are N-1 1s where N is 8 -
(BIT_COUNT % 8) and then a 0 so the total bits at the start equal to N
which is also the number of the unwanted trailing bits at the end.